remove dependence on config in app and script files + refactor site functions to data_utils
