+++ scripts/get_latest_baselines.py
@@ -9,17 +9,18 @@ Copyright (C) 2021-2023 Government of Canada
 Authors:
   - Drew Rotheram <drew.rotheram-clarke@nrcan-rncan.gc.ca>
 """
-import configparser
 import os
 import botocore.exceptions
 import boto3
 import yaml
 
+from app.data_utils import get_config_params
+
 
 def main():
     '''Main function, get latest perpendicular
     baseline files for all site/beam combos'''
-    config = get_config_params("config.ini")
+    config = get_config_params()
     s_3 = boto3.client('s3')
 
     with open('app/Data/beamList.yml', encoding="utf-8") as beam_list_yml:
@@ -31,20 +32,11 @@ def main():
             if not os.path.exists(f'app/Data/{site}/{beam}'):
                 os.makedirs(f'app/Data/{site}/{beam}')
             try:
-                s_3.download_file(Bucket=config.get('AWS', 'bucketName'),
+                s_3.download_file(Bucket=config['AWS_BUCKET_NAME'],
                                   Key=f'{site}/{beam}/bperp_all',
                                   Filename=f'app/Data/{site}/{beam}/bperp_all')
             except botocore.exceptions.ClientError:
                 print('Perpendicular Baseline File not found')
-
-
-def get_config_params(args):
-    """
-    Parse command line arguments
-    """
-    config_parse_obj = configparser.ConfigParser()
-    config_parse_obj.read(args)
-    return config_parse_obj
 
 
 if __name__ == '__main__':
